修改错误文件夹导致没打入资源文件。

+++ zt-example-dao/src/main/resources/database/example-hsqldb-schema.sql
@@ -3,7 +3,7 @@
 create table example_user (
   user_id       varchar(50)    not null,
   password      varchar(50)    not null,
-  status        number(1)   not null ,
+  status        int   not null ,
   create_date   date     not null,
   constraint pk_item1 primary key (user_id)
 );
